requirements.txt
PyQt5>=5.15.0
PyOpenGL>=3.1.0
numpy>=1.19.0
pybullet>=3.2.0
pygame>=2.0.0 